Boardコンポーネントについて格納フォルダのタイポを修正。状態の管理をstateからreduxへ
